Implement concurrent execution of client requests
Currently, the server's execution limits are not checked,
and clients are not informed if their request has been received,
has begun processing, or has concluded.

Furthermore, there is still no handling of priorities for different
requests, but this commit will make all of the above possible
by creating a `Monitor` struct that can be updated to do all these things.

The idea is to have an `mpsc::channel` with the server as consumer,
and each monitor as producer, and insert tasks in a priority queue
stored in the server's main thread.

It is still unclear whether the `Monitor` should be in charge of
communicating with the client that made the request or if the server
should do it, as the former will entail wrapping the server's `UdSocket`
and pass it to every `Monitor`'s thread.